% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption_cwb.R
\name{add_caption_cwb}
\alias{add_caption_cwb}
\title{add_caption_cwb}
\usage{
add_caption_cwb(
  type = c("plot", "table"),
  include_data_source = FALSE,
  data_source = NULL,
  github_username = "colebaril",
  bluesky_username = "@colebaril.ca"
)
}
\arguments{
\item{type}{Either "plot" or "table"}

\item{include_data_source}{Boolean; whether to include data source}

\item{data_source}{The source of the data}

\item{github_username}{Your GitHub username (default: "colebaril")}

\item{bluesky_username}{Your Bluesky username}

\item{x}{A ggplot object or gt table}
}
\value{
A ggplot or gt object with a caption applied
}
\description{
Add a social media caption to a ggplot or gt table.
}
