% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commas.R
\name{add_commas}
\alias{add_commas}
\title{Adds commas to large numbers for easier readability.}
\usage{
add_commas(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A character vector of the same length as \code{x} with commas inserted as thousand separators.
}
\description{
Adds commas to large numbers for easier readability.
}
