% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{extract_table_reader}
\alias{extract_table_reader}
\title{Adapter reader for extract_table()}
\usage{
extract_table_reader(file, ...)
}
\description{
Allows extract_table() to be used with read_data_tree()
}
