% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table_tree.R
\name{read_table_tree}
\alias{read_table_tree}
\title{Read extracted tables from an Excel directory tree}
\usage{
read_table_tree(path, start_column, ext = "xlsx", recursive = TRUE, ...)
}
\description{
Read extracted tables from an Excel directory tree
}
