% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption_cwb.R
\name{social_caption}
\alias{social_caption}
\title{Generate social caption for ggplot or gt table}
\usage{
social_caption(
  include_data_source = FALSE,
  data_source = NULL,
  github_username = "colebaril",
  bluesky_username = "@colebaril.ca"
)
}
\arguments{
\item{include_data_source}{Logical; whether to include data source}

\item{data_source}{Character; the source of the data}

\item{github_username}{Character; GitHub username (default: "colebaril")}

\item{bluesky_username}{Character; Bluesky username (default: "@colebaril.ca")}
}
\value{
Character string containing HTML caption
}
\description{
Returns a caption string with optional data source and social icons,
suitable for passing directly to labs(caption = …) in ggplot.
}
