% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spooky.R
\name{theme_spooky}
\alias{theme_spooky}
\title{theme_spooky()}
\usage{
theme_spooky()
}
\value{
A \code{ggplot2} theme object.
}
\description{
A minimalist spooky theme for ggplot2.
Light background, subtle grid lines, and muted accents.
}
\examples{
library(ggplot2)
library(cwb)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_spooky()

}
