% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_theme.R
\name{trashpanda_quarto_theme}
\alias{trashpanda_quarto_theme}
\title{trashpanda Quarto revealjs themes}
\usage{
trashpanda_quarto_theme(name = c("dark", "light", "serif"))
}
\arguments{
\item{name}{Theme name: "dark", "light"}
}
\description{
trashpanda Quarto revealjs themes
}
