% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quarto_theme.R
\name{load_quarto_theme}
\alias{load_quarto_theme}
\title{Load a Trashpanda Quarto theme into the current project}
\usage{
load_quarto_theme(
  name = c("dark", "light"),
  dest_folder = "styles",
  overwrite = FALSE
)
}
\arguments{
\item{name}{Theme name ("dark", "light")}

\item{dest_folder}{Destination folder in project (default: "styles")}

\item{overwrite}{Whether to overwrite if the CSS already exists (default: FALSE)}
}
\value{
Relative path to the CSS file in your project
}
\description{
Copies a precompiled CSS theme into the project 'styles/' folder for use in Quarto.
}
