% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cole.R
\name{theme_cole}
\alias{theme_cole}
\title{theme_cole()}
\usage{
theme_cole(
  base_size = 12,
  base_family = "sans",
  remove_grid = FALSE,
  show_axis_lines = c("bottom", "left"),
  dark = FALSE
)
}
\arguments{
\item{base_size}{Base text size. Default 12.}

\item{base_family}{Base font family. Default "sans".}

\item{remove_grid}{Logical. If TRUE, removes all grid lines.}

\item{show_axis_lines}{Shows axis lines, text, titles and ticks. c("all", "bottom", "top", "right", "left"); default:
c("bottom", "left").}

\item{dark}{Logical. If TRUE, plot is transformed to a dark theme.}
}
\value{
A \code{ggplot2} theme object that can be added to ggplot plots.
}
\description{
A \code{ggplot2} theme styled to resemble old parchment and ink, giving plots
a vintage, manuscript-like appearance.
}
\details{
This theme adjusts panel backgrounds, grid lines, and text colors to
evoke the look of old parchment and handwritten ink. Works with \code{ggplot2} plots.
}
\examples{
library(ggplot2)
library(trashpanda)
library(palmerpenguins)

ggplot(penguins, aes(flipper_length_mm, bill_length_mm, fill = species, group = species)) +
  geom_point(shape = 21) +
  labs(title = "Flipper Length vs. Bill Length",
       subtitle = "Lorem ipsum") +
  theme_cole(show_axis_lines = c("bottom", "left"), remove_grid = TRUE, dark = TRUE) +
  add_caption_cwb() 
  
  
ggplot(penguins, aes(flipper_length_mm, bill_length_mm, fill = species, group = species)) +
  geom_point(shape = 21) +
  labs(title = "Flipper Length vs. Bill Length",
       subtitle = "Lorem ipsum") +
  theme_cole(show_axis_lines = c("bottom", "left"), remove_grid = FALSE) +
  add_caption_cwb() 

}
